refactor: ToolTipCards -> EntryToolTipCard
