+++ VocaDbWeb/New/components/ToolTips/EntryToolTip.tsx
@@ -1,18 +1,25 @@
 import React from 'react';
-import { AlbumToolTipProps } from './AlbumToolTipCard';
-import { ArtistToolTipProps } from './ArtistToolTipCard';
+import { AlbumToolTipProps } from './AlbumToolTipContent';
+import { ArtistToolTipProps } from './ArtistToolTipContent';
 
-const ArtistToolTipCard = React.lazy(() => import('./ArtistToolTipCard'));
-const AlbumToolTipCard = React.lazy(() => import('./AlbumToolTipCard'));
+const EntryToolTipCard = React.lazy(() => import('./EntryToolTipCard'));
+const ArtistToolTipContent = React.lazy(() => import('./ArtistToolTipContent'));
+const AlbumToolTipCard = React.lazy(() => import('./AlbumToolTipContent'));
 
-type EntryToolTipProps = AlbumToolTipProps | ArtistToolTipProps;
+type EntryToolTipProps = (AlbumToolTipProps | ArtistToolTipProps) & { children: JSX.Element };
 
 export default function EntryToolTip(props: EntryToolTipProps) {
-	if (props.entry === 'album') {
-		return <AlbumToolTipCard {...props} />;
-	} else if (props.entry === 'artist') {
-		return <ArtistToolTipCard {...props} />;
-	}
-	return <></>;
+	const getToolTip = () => {
+		if (props.entry === 'album') {
+			return <AlbumToolTipCard {...props} />;
+		} else if (props.entry === 'artist') {
+			return <ArtistToolTipContent {...props} />;
+		}
+	};
+	return (
+		<React.Suspense fallback={props.children}>
+			<EntryToolTipCard tooltip={getToolTip()}>{props.children}</EntryToolTipCard>
+		</React.Suspense>
+	);
 }
