+++ VocaDbWeb/New/components/ToolTips/EntryToolTipCard.tsx
@@ -1,0 +1,26 @@
+import { Paper, Tooltip } from '@mantine/core';
+
+interface EntryToolTipCardProps {
+	children: JSX.Element;
+	tooltip?: JSX.Element;
+}
+
+export default function EntryToolTipCard({ children, tooltip }: EntryToolTipCardProps) {
+	return (
+		<Tooltip
+			styles={{
+				tooltip: {
+					padding: 0,
+				},
+			}}
+			label={
+				<Paper radius="xs" p="md">
+					{tooltip}
+				</Paper>
+			}
+		>
+			{children}
+		</Tooltip>
+	);
+}
+
